Introduce spring-data-rest for endpoints
