+++ backend-core/build.gradle
@@ -1,0 +1,11 @@
+dependencies {
+    compile libraries.application.springRestData
+
+    compile libraries.persistence.jpaApi
+    compile libraries.persistence.hibernateCore
+    compile libraries.persistence.springDataJpa
+
+    compile libraries.commons.apacheLang3
+    compile libraries.commons.nvI18n
+    compile libraries.commons.beanValidation
+}
