+++ backend-core/src/main/java/eu/socialedge/hermes/backend/core/ext/Identifiable.java
@@ -1,0 +1,42 @@
+/*
+ * Hermes - The Municipal Transport Timetable System
+ * Copyright (c) 2017 SocialEdge
+ * <p>
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ * <p>
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ */
+package eu.socialedge.hermes.backend.core.ext;
+
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.experimental.Accessors;
+
+import javax.persistence.*;
+import java.io.Serializable;
+
+import static org.apache.commons.lang3.Validate.notNull;
+
+@Getter
+@EqualsAndHashCode
+@Accessors(fluent = true)
+@NoArgsConstructor(force = true)
+@MappedSuperclass @Access(AccessType.FIELD)
+public class Identifiable<T> implements Serializable {
+
+    @Id
+    @Column(name = "id")
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    protected final T id;
+
+    public Identifiable(T id) {
+        this.id = notNull(id);
+    }
+}
