+++ backend-core/src/main/java/eu/socialedge/hermes/backend/core/repository/AgencyRepository.java
@@ -1,0 +1,25 @@
+/*
+ * Hermes - The Municipal Transport Timetable System
+ * Copyright (c) 2017 SocialEdge
+ * <p>
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ * <p>
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ */
+package eu.socialedge.hermes.backend.core.repository;
+
+import eu.socialedge.hermes.backend.core.Agency;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.data.rest.core.annotation.RepositoryRestResource;
+import org.springframework.stereotype.Repository;
+
+@Repository
+@RepositoryRestResource
+public interface AgencyRepository extends CrudRepository<Agency, Long> {
+}
