--- backend-domain/src/main/java/eu/socialedge/hermes/backend/domain/repository/TripRepository.java
@@ -1,21 +1,0 @@
-/*
- * Hermes - The Municipal Transport Timetable System
- * Copyright (c) 2017 SocialEdge
- * <p>
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- * <p>
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- */
-package eu.socialedge.hermes.backend.domain.repository;
-
-import eu.socialedge.hermes.backend.domain.Trip;
-import org.springframework.data.repository.CrudRepository;
-
-public interface TripRepository extends CrudRepository<Trip, Long> {
-}
